+++ rust/tests/lc_x448_test.rs
@@ -17,57 +17,57 @@
  * DAMAGE.
  */
 
-use leancrypto_sys::lcr_x25519::lcr_x25519;
+use leancrypto_sys::lcr_x448::lcr_x448;
 use wycheproof::{
     xdh::{TestName, TestSet},
     TestResult,
 };
 
-fn lc_rust_x25519_one() {
-    let mut x25519_local = lcr_x25519::new();
-    let mut x25519_remote = lcr_x25519::new();
+fn lc_rust_x448_one() {
+    let mut x448_local = lcr_x448::new();
+    let mut x448_remote = lcr_x448::new();
 
-    let result = x25519_local.enable();
+    let result = x448_local.enable();
     assert_eq!(result, Ok(()));
 
     /* Generate local key pair */
-    let result = x25519_local.keypair();
+    let result = x448_local.keypair();
     assert_eq!(result, Ok(()));
 
     /* Export local public key */
-    let pk_local_slice = x25519_local.get_pk().expect("get_pk");
+    let pk_local_slice = x448_local.get_pk().expect("get_pk");
     let pk_local = pk_local_slice.to_vec();
 
     /* Generate remote key pair */
-    let result = x25519_remote.keypair();
+    let result = x448_remote.keypair();
     assert_eq!(result, Ok(()));
 
     /* Export remote public key */
-    let pk_remote_slice = x25519_remote.get_pk().expect("get_pk");
+    let pk_remote_slice = x448_remote.get_pk().expect("get_pk");
     let pk_remote = pk_remote_slice.to_vec();
 
     /* Load remote PK into local context */
-    let result = x25519_local.pk_remote_load(&pk_remote);
+    let result = x448_local.pk_remote_load(&pk_remote);
     assert_eq!(result, Ok(()));
 
     /* Load local PK into remote context */
-    let result = x25519_remote.pk_remote_load(&pk_local);
+    let result = x448_remote.pk_remote_load(&pk_local);
     assert_eq!(result, Ok(()));
 
     /* Generate local shared secret */
-    let result = x25519_local.shared_secret();
+    let result = x448_local.shared_secret();
     assert_eq!(result, Ok(()));
 
     /* Export local shared secret */
-    let ss_local_slice = x25519_local.get_ss().expect("get_ss");
+    let ss_local_slice = x448_local.get_ss().expect("get_ss");
     let ss_local = ss_local_slice.to_vec();
 
     /* Generate remote shared secret */
-    let result = x25519_remote.shared_secret();
+    let result = x448_remote.shared_secret();
     assert_eq!(result, Ok(()));
 
     /* Export remote shared secret */
-    let ss_remote_slice = x25519_remote.get_ss().expect("get_ss");
+    let ss_remote_slice = x448_remote.get_ss().expect("get_ss");
     assert_eq!(result, Ok(()));
     let ss_remote = ss_remote_slice.to_vec();
 
@@ -78,34 +78,35 @@ fn lc_rust_x25519_one() {
 }
 
 #[test]
-fn lc_rust_x25519() {
-    lc_rust_x25519_one();
+fn lc_rust_x448() {
+    lc_rust_x448_one();
 }
 
 #[test]
-fn x25519() {
-    let test_set = TestSet::load(TestName::X25519).unwrap();
+fn x448() {
+    let test_set = TestSet::load(TestName::X448).unwrap();
     for test_group in &test_set.test_groups {
         for test in &test_group.tests {
             println!("Test case {}: {}", test.tc_id, test.comment);
 
-            let mut x25519 = leancrypto_sys::lcr_x25519::lcr_x25519::new();
-            let result = x25519.enable();
+            let mut x448 = leancrypto_sys::lcr_x448::lcr_x448::new();
+            let result = x448.enable();
             assert_eq!(result, Ok(()));
 
-            let result = x25519.sk_load(&test.private_key);
+            let result = x448.sk_load(&test.private_key);
             assert_eq!(result, Ok(()));
-            let result = x25519.pk_remote_load(&test.public_key);
-            assert_eq!(result, Ok(()));
+            let mut result = x448.pk_remote_load(&test.public_key);
 
-            let result = x25519.shared_secret();
+            if result == Ok(()) {
+                result = x448.shared_secret();
+            }
 
             match &test.result {
                 TestResult::Invalid => {
                     assert!(result.is_err());
                 }
                 TestResult::Valid | TestResult::Acceptable => {
-                    let ss_slice = x25519.get_ss().expect("get_ss");
+                    let ss_slice = x448.get_ss().expect("get_ss");
                     assert_eq!(result, Ok(()));
                     assert_eq!(
                         ss_slice[..],